Updated for 5.2. This version requires GridStatusRaidDebuff MoP 5.2/r11 or higher.
This will generate a popup warning if you are running an older incompatible
version. I appologize for the pop-up again, but it is better than things
being broken and getting a lot of bug reports.

This change was required because libbabble-zone is deprecated in 5.2 and I
had to rewrite GridStatusRaidDebuff MoP to no longer require it for new
dungeons.  Libbabble-zone is still being included in GridStatusRaidDebuff
for old raid packs that I do not have access to update so they remain
functional as long as possible.